+++ src/classes/ui/buttons/ToolBtn.ts
@@ -1,0 +1,137 @@
+import { Game } from "@/scenes/Game";
+import { Tools } from "@/types";
+import { getLocal, availableTools } from "@/utils";
+import { BaseBtn } from "./BaseBtn";
+
+export class ToolBtn extends BaseBtn {
+  scene: Game;
+
+  icon: Phaser.GameObjects.Image;
+  currentCountText: Phaser.GameObjects.Text;
+
+  toolKey: Tools;
+  localText: string;
+
+  initialCount: number;
+  currentCount: number;
+
+  isSelected: boolean = false;
+
+  constructor(
+    scene: Game,
+    x: number,
+    y: number,
+    count: number,
+    hotkey: string,
+    toolKey: Tools
+  ) {
+    super(scene, x, y);
+
+    const {
+      game: { ui },
+    } = getLocal(scene);
+
+    this.scene = scene;
+    this.toolKey = toolKey;
+    this.localText = ui.toolUses;
+    this.initialCount = count;
+    this.currentCount = count;
+
+    const { textureKey, props } =
+      availableTools[toolKey as keyof typeof availableTools];
+
+    this.icon = scene.make
+      .image({ key: textureKey, ...props })
+      .setScale(0.6)
+      .setAlpha(0.8);
+
+    this.currentCountText = scene.make
+      .text({
+        x: 0,
+        y: -65,
+        text: `${this.localText}: ${this.currentCount}`,
+        style: { font: "24px OpenSans_ExtraBold" },
+      })
+      .setOrigin(0.5);
+
+    const hotkeyBtn = scene.make.image({
+      key: "uiatlas",
+      frame: "hotkey_btn",
+      scale: 0.3,
+      y: 50,
+    });
+
+    const hotkeyText = scene.make
+      .text({
+        text: hotkey,
+        x: hotkeyBtn.x,
+        y: hotkeyBtn.y,
+        style: {
+          color: "#3e3e3e",
+          font: "24px OpenSans_Bold",
+        },
+      })
+      .setOrigin(0.5);
+
+    this.container.add([
+      this.icon,
+      this.currentCountText,
+      hotkeyBtn,
+      hotkeyText,
+    ]);
+
+    const keyObj = scene.input.keyboard?.addKey(hotkey);
+    if (keyObj) keyObj.on("down", this.onClick, this);
+
+    this.setInteractive(
+      () => this.onClick(),
+      () => this.currentCount > 0
+    );
+  }
+
+  onClick() {
+    if (this.scene.grid.activeSwap) {
+      this.scene.grid.activeSwap.cancelSwapSelection();
+      this.scene.changeSelectedTool(Tools.none);
+      return;
+    }
+
+    if (this.currentCount === 0) return;
+
+    this.isSelected ? this.deselect() : this.select();
+  }
+
+  select() {
+    this.scene.sound.play("colorSelect");
+    this.scene.changeSelectedTool(this.toolKey);
+    this.isSelected = true;
+    this.icon.setTintFill(0xffcd3f);
+    this.setSelected(true);
+  }
+
+  deselect() {
+    this.scene.sound.play("colorSelect", { detune: 100 });
+    this.scene.changeSelectedTool(Tools.none);
+    this.isSelected = false;
+    this.icon.clearTint();
+    this.setSelected(false);
+  }
+
+  update() {
+    this.isSelected = this.scene.gameStates.selectedTool === this.toolKey;
+    this.setSelected(this.isSelected);
+    if (!this.isSelected) this.icon.clearTint();
+
+    this.currentCountText.setText(`${this.localText}: ${this.currentCount}`);
+  }
+
+  reset() {
+    this.currentCount = this.initialCount;
+    this.update();
+  }
+
+  decrease() {
+    this.currentCount--;
+    this.update();
+  }
+}
